// nix/modules/container-services/bewcloud/lib/types.ts
export type PartialDeep<T> =
  | (T extends (infer U)[]
      ? PartialDeep<U>[]
      : { [P in keyof T]?: PartialDeep<T[P]> })
  | T;

export type OptionalApp = "news" | "notes" | "photos" | "expenses" | "contacts";

export interface Config {
  auth: {
    /** The base URL of the application you use to access the app, i.e. "http://localhost:8000" or "https://cloud.example.com" */
    baseUrl: string;
    /** If true, anyone can sign up for an account. Note that it's always possible to sign up for the first user, and they will be an admin */
    allowSignups: boolean;
    /** If true, email verification will be required for signups (using SMTP settings below) */
    enableEmailVerification: boolean;
    /** If true, all signups become active for 100 years */
    enableForeverSignup: boolean;
    /** If true, users can enable multi-factor authentication (TOTP, Passkeys, or Email if the SMTP settings below are set) */
    enableMultiFactor: boolean;
    /** Can be set to allow more than the baseUrl's domain for session cookies */
    allowedCookieDomains: string[];
    /** If true, the cookie domain will not be strictly set and checked against. This skipping slightly reduces security, but is usually necessary for reverse proxies like Cloudflare Tunnel. */
    skipCookieDomainSecurity: boolean;
    /** If true, single sign-on will be enabled */
    enableSingleSignOn: boolean;
    /** The Discovery URL (AKA Issuer) of the identity/single sign-on provider */
    singleSignOnUrl: string;
    /** The attribute to prefer as email of the identity/single sign-on provider */
    singleSignOnEmailAttribute: string;
    /** The scopes to request from the identity/single sign-on provider */
    singleSignOnScopes: string[];
  };
  files: {
    /** The root-relative root path for files, i.e. "data-files" */
    rootPath: string;
    /** If true, public file sharing will be allowed (still requires a user to enable sharing for a given file or directory) */
    allowPublicSharing: boolean;
  };
  core: {
    /** dashboard and files cannot be disabled */
    enabledApps: OptionalApp[];
  };
  visuals: {
    /** An override title of the application. Empty shows the default title. */
    title: string;
    /** An override description of the application. Empty shows the default description. */
    description: string;
    /** The email address to contact for help. Empty will disable/hide the "need help" sections. */
    helpEmail: string;
  };
  email: {
    /** The email address to send emails from */
    from: string;
    /** The SMTP host to send emails from */
    host: string;
    /** The SMTP port to send emails from */
    port: number;
  };
  contacts: {
    /** If true, the CardDAV server will be enabled (proxied) */
    enableCardDavServer: boolean;
    /** The CardDAV server URL to proxy to */
    cardDavUrl: string;
  };
  calendar: {
    /** If true, the CalDAV server will be enabled (proxied) */
    enableCalDavServer: boolean;
    /** The CalDAV server URL to proxy to */
    calDavUrl: string;
  };
}